backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
openai==1.3.8
anthropic==0.7.8
google-generativeai==0.3.2
psycopg2-binary==2.9.9
supabase==2.0.2
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
nltk==3.8.1
textblob==0.17.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==1.5.3